+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `Auto Piercing` OP ability (requires TabletopTweaks-Base).
 - Added UMM option to allow for multiple Mythic OP abilities and Mythic special powers.
 - Updated mod support for ExpandedContent.
+### Changed
+- Reworked UMM settings.
 ### Fixed
 - Fixed missing dialogue for drow ambush encounter.
 - Fixed `Channel Energy` and other features class level scaling in legacy progressions.